third_party/blink/web_tests/platform/generic/editing/pasteboard/paste-text-with-style-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification

Markup before:
| "
<#selection-caret>"
| <b>
|   <i>
|     "hello bold and italic"
| <div>
|   <i>
|     "hello italic"
| <div>
|   <font>
|     color="#ff0000"
|     "hello red"
| "
"

Markup after:
| <b>
|   <i>
|     "hello bold and italic"
| <div>
|   <i>
|     "hello italic"
| <div>
|   <font>
|     color="#ff0000"
|     "hello red<#selection-caret>"
| "
"
